fix(recovery): Finding 06 — widen uint8_t counts[256] to uint32_t
Two voting/averaging functions used uint8_t counters for byte-value
frequency tables that are driven by size_t / uint8_t pass-count loops.
The 256th identical input wraps the counter back to zero and turns
the majority winner into a near-loser without warning.

Sites:
  src/recovery/uft_sector_recovery.c:95   average_sector_reads()
  src/recovery/uft_multiread_pipeline.c:222  vote_byte()

The bug is dormant in production today because:
  * MULTIREAD_MAX_PASSES is 16
  * average_sector_reads() is reached with read_count + 1 ≤ 17

But the function signatures both accept the wider type
(read_count is size_t, pass_count is uint8_t with room to 255), so
any future stress harness, external recovery wizard or bump of the
MAX_PASSES constant trips the wrap silently. Forensic recovery
cannot tolerate a silent vote-flip.

Changes
-------
* `uint8_t counts[256]` → `uint32_t counts[256]` in both files.
* Local `best_count` / `max_count` variables widened to match.
* Confidence formulas reworked to use unsigned 32-bit arithmetic
  (`* 255u`, `* 100u`) so the new wider type doesn't surprise
  signed-int promotion. Result still narrows to uint8_t at the
  output boundary; range remains 0..255.
* Added explanatory comments above each counts[] declaration so
  the next reader knows why the type matters.

Verification
------------
gcc -fsyntax-only -Wall -Wextra clean (mingw 13.1.0).
uft-recovery-integrity/lint_recovery.sh on both files: clean — all
five invariants pass for both source files for the first time.

Refs: HTML logic audit Finding 06 (MEDIUM); recovery-integrity audit
Bugs 4 and 5 (I4 family).

+++ src/recovery/uft_multiread_pipeline.c
@@ -218,9 +218,14 @@ static uint8_t vote_byte(const multiread_pass_t *passes, uint8_t pass_count,
     }
 
     /* Count occurrences of each value, restricted to authoritative
-     * passes when at least one such pass is available. */
-    uint8_t counts[256] = {0};
-    uint8_t valid_count = 0;
+     * passes when at least one such pass is available.
+     *
+     * counts[] and max_count widened to uint32_t (Finding 06):
+     * pass_count is uint8_t today (max 255) so the bug is dormant,
+     * but a future MULTIREAD_MAX_PASSES bump or a signature widening
+     * would silently overflow a uint8_t counter. Defense in depth. */
+    uint32_t counts[256] = {0};
+    uint8_t  valid_count = 0;
 
     for (uint8_t p = 0; p < pass_count; p++) {
         if (only_crc_ok && !passes[p].crc_ok) continue;
@@ -237,19 +242,19 @@ static uint8_t vote_byte(const multiread_pass_t *passes, uint8_t pass_count,
     }
 
     /* Find most common value */
-    uint8_t max_count = 0;
-    uint8_t max_val = 0;
+    uint32_t max_count = 0;
+    uint8_t  max_val   = 0;
 
     for (int v = 0; v < 256; v++) {
         if (counts[v] > max_count) {
             max_count = counts[v];
-            max_val = (uint8_t)v;
+            max_val   = (uint8_t)v;
         }
     }
 
     /* Calculate confidence */
     if (confidence) {
-        *confidence = (uint8_t)((max_count * 100) / valid_count);
+        *confidence = (uint8_t)((max_count * 100u) / valid_count);
     }
 
     /* Detect weak bits (not unanimous) */
